+++ lib/src/deprecated/302.dart
@@ -1,0 +1,13 @@
+// Project imports:
+import 'package:zego_uikit_prebuilt_live_streaming/zego_uikit_prebuilt_live_streaming.dart';
+
+const deprecatedTipsV302 = ', '
+    'deprecated since 3.0.2, '
+    'will be removed after 3.1.0';
+
+@Deprecated('Use ZegoLiveStreamingPKMixerLayout instead$deprecatedTipsV302')
+typedef ZegoPKMixerLayout = ZegoLiveStreamingPKMixerLayout;
+
+@Deprecated(
+    'Use ZegoLiveStreamingPKMixerDefaultLayout instead$deprecatedTipsV302')
+typedef ZegoPKMixerDefaultLayout = ZegoLiveStreamingPKMixerDefaultLayout;