+++ lib/src/pk/layout/layout.dart
@@ -6,21 +6,21 @@ import 'package:flutter/services.dart';
 import 'package:zego_uikit_prebuilt_live_streaming/src/pk/layout/prefer_gird_layout.dart';
 
 /// width of mixer canvas
-const zegoPK2MixerCanvasWidth = 864.0;
+const zegoLiveStreamingPK2MixerCanvasWidth = 972.0;
 
 /// height of mixer canvas
-const zegoPK2MixerCanvasHeight = 972.0;
+const zegoLiveStreamingPK2MixerCanvasHeight = 864.0;
 
 /// Inheritance of the hybrid layout parent class allows you to return your
 /// custom coordinates and modify the layout of the mixed stream.
-/// You can refer to [ZegoPKPreferGridMixerLayout] or [ZegoPKV2GridMixerLayout].
-abstract class ZegoPKMixerLayout {
+/// You can refer to [ZegoLiveStreamingPKPreferGridMixerLayout] or [ZegoPKV2GridMixerLayout].
+abstract class ZegoLiveStreamingPKMixerLayout {
   /// The size of the mixed stream canvas.
   /// default is zegoPK2MixerCanvasWidth x zegoPK2MixerCanvasHeight
   Size getResolution() {
     return const Size(
-      zegoPK2MixerCanvasWidth,
-      zegoPK2MixerCanvasHeight,
+      zegoLiveStreamingPK2MixerCanvasWidth,
+      zegoLiveStreamingPK2MixerCanvasHeight,
     );
   }
 
@@ -31,4 +31,5 @@ abstract class ZegoPKMixerLayout {
   });
 }
 
-typedef ZegoPKMixerDefaultLayout = ZegoPKPreferGridMixerLayout;
+typedef ZegoLiveStreamingPKMixerDefaultLayout
+    = ZegoLiveStreamingPKPreferGridMixerLayout;